// Frontend/src/app/api/models/application-detail-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ApplicationDetailDto {
  address?: string;
  candidateId?: number;
  cvUrl?: string;
  email?: string;
  evaluationCount?: number;
  firstName?: string;
  highestDegree?: string;
  id?: number;
  jobOfferId?: number;
  jobOfferTitle?: string;
  justificationIA?: string;
  lastName?: string;
  majorField?: string;
  matchingCompetences?: number;
  matchingDiploma?: number;
  matchingExperience?: number;
  scoreIA?: number;
  status?: 'DRAFT' | 'SUBMITTED' | 'UNDER_REVIEW' | 'AI_SCORED' | 'SHORTLISTED' | 'INTERVIEW_SCHEDULED' | 'INTERVIEW_COMPLETED' | 'ACCEPTED' | 'REJECTED' | 'WITHDRAWN';
  submitDate?: string;
  updateDate?: string;
}